refactor attendance
