+++ src/components/calendar/Attendance.vue
@@ -72,7 +72,7 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue';
 import Card from '@components/card.vue';
-import AttendanceCell from '@components/AttendanceCell.vue';
+import AttendanceCell from '@/components/calendar/AttendanceCell.vue';
 
 const props = defineProps<{
   initialYear?: number;
